sentinel-ai/requirements.txt
# Sentinel AI Dependencies
# =========================

# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Satellite Data
earthengine-api>=0.1.370
geemap>=0.30.0

# Image Processing
rasterio>=1.3.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
